feat: migrate /phoneAlert endpoint to AlertInfoController
- Migrated getPhoneNumbersByStation method from FireStationController to AlertInfoController
- Moved related business logic from FireStationService to AlertInfoService
- Added @Autowired dependency for FireStationRepository in AlertInfoService
- Verified endpoint functionality at /phoneAlert?firestation=<stationNumber> via Postman
